+++ example/lib/main.dart
@@ -1,3 +1,5 @@
+import 'dart:io';
+
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
@@ -31,67 +33,86 @@ class _MyAppState extends State<MyApp> {
       theme: ThemeData.light(),
       darkTheme: ThemeData.dark(),
       themeMode: ThemeMode.system,
-      home: Scaffold(
-        appBar: PreferredSize(
-          preferredSize: Size.fromHeight(kToolbarHeight),
-          child: Stack(
-            children: <Widget>[
-              Positioned.fill(
-                child: GestureDetector(
-                  onTapDown: (_) {
-                    WindowUtils.startDrag();
-                  },
-                  onDoubleTap: () {
-                    WindowUtils.windowTitleDoubleTap();
-                  },
-                  child: Material(
-                    elevation: 4.0,
-                    color: Theme.of(context).primaryColor,
+      home: WindowsFrame(
+        active: Platform.isWindows,
+        border: Border.all(color: Colors.grey),
+        child: Scaffold(
+          appBar: PreferredSize(
+            preferredSize: Size.fromHeight(kToolbarHeight),
+            child: Stack(
+              children: <Widget>[
+                Positioned.fill(
+                  child: GestureDetector(
+                    onTapDown: (_) {
+                      WindowUtils.startDrag();
+                    },
+                    onDoubleTap: () {
+                      WindowUtils.windowTitleDoubleTap().then((_) {
+                        if (mounted) setState(() {});
+                      });
+                    },
+                    child: Material(
+                      elevation: 4.0,
+                      color: Theme.of(context).primaryColor,
+                    ),
                   ),
                 ),
-              ),
-              Positioned.fill(
-                child: IgnorePointer(
-                  child: Center(
-                    child: Text(
-                      'Window Utils Example',
-                      style: TextStyle(
-                        color: Colors.white,
-                        fontSize: 20.0,
+                Positioned.fill(
+                  child: IgnorePointer(
+                    child: Center(
+                      child: Text(
+                        'Window Utils Example',
+                        style: TextStyle(
+                          color: Colors.white,
+                          fontSize: 20.0,
+                        ),
                       ),
                     ),
                   ),
                 ),
-              ),
-              Positioned(
-                bottom: 0,
-                top: 0,
-                right: 0,
-                child: Center(
-                  child: Row(
-                    children: <Widget>[
-                      IconButton(
-                        color: Colors.white,
-                        icon: Icon(Icons.info_outline),
-                        onPressed: () {
-                          WindowUtils.getWindowSize()
-                              .then((val) => print('Window: $val'));
-                          WindowUtils.getScreenSize()
-                              .then((val) => print('Screen: $val'));
-                          WindowUtils.getWindowOffset()
-                              .then((val) => print('Offset: $val'));
-                        },
-                      ),
-                    ],
+                Positioned(
+                  bottom: 0,
+                  top: 0,
+                  right: 0,
+                  child: Center(
+                    child: Row(
+                      children: <Widget>[
+                        IconButton(
+                          color: Colors.white,
+                          icon: Icon(Icons.info_outline),
+                          onPressed: () {
+                            WindowUtils.getWindowSize()
+                                .then((val) => print('Window: $val'));
+                            WindowUtils.getScreenSize()
+                                .then((val) => print('Screen: $val'));
+                            WindowUtils.getWindowOffset()
+                                .then((val) => print('Offset: $val'));
+                          },
+                        ),
+                      ],
+                    ),
                   ),
                 ),
-              ),
-            ],
+              ],
+            ),
           ),
-        ),
-        body: Container(
-          child: ListView(
+          floatingActionButton: InkWell(
+            child: Icon(Icons.drag_handle),
+          ),
+          body: ListView(
             children: <Widget>[
+              ListTile(
+                title: Text("Max Window Size"),
+                trailing: IconButton(
+                  icon: Icon(Icons.desktop_windows),
+                  onPressed: () {
+                    WindowUtils.getScreenSize().then((val) async {
+                      await WindowUtils.setSize(Size(val.width, val.height));
+                      await WindowUtils.setPosition(Offset(0, 0));
+                    });
+                  },
+                ),
+              ),
               ListTile(
                 title: Text("Increase Window Size"),
                 trailing: IconButton(
@@ -139,6 +160,112 @@ class _MyAppState extends State<MyApp> {
             ],
           ),
         ),
+      ),
+    );
+  }
+}
+
+class WindowsFrame extends StatelessWidget {
+  final Widget child;
+  final bool active;
+  final BoxBorder border;
+
+  const WindowsFrame({
+    Key key,
+    @required this.child,
+    @required this.active,
+    this.border,
+  }) : super(key: key);
+  @override
+  Widget build(BuildContext context) {
+    if (!active) return child;
+    final width = 4.0;
+    final height = 4.0;
+    return Container(
+      child: Stack(
+        fit: StackFit.expand,
+        children: <Widget>[
+          border != null
+              ? Container(
+                  decoration: BoxDecoration(border: border),
+                  child: child,
+                )
+              : child,
+          Positioned(
+            top: 0,
+            left: width,
+            right: width,
+            height: height,
+            child: GestureDetector(
+              onTapDown: (_) => WindowUtils.startResize(DragPosition.top),
+            ),
+          ),
+          Positioned(
+            bottom: 0,
+            left: width,
+            right: width,
+            height: height,
+            child: GestureDetector(
+              onTapDown: (_) => WindowUtils.startResize(DragPosition.bottom),
+            ),
+          ),
+          Positioned(
+            bottom: height,
+            top: height,
+            right: 0,
+            width: width,
+            child: GestureDetector(
+              onTapDown: (_) => WindowUtils.startResize(DragPosition.right),
+            ),
+          ),
+          Positioned(
+            bottom: height,
+            top: height,
+            left: 0,
+            width: width,
+            child: GestureDetector(
+              onTapDown: (_) => WindowUtils.startResize(DragPosition.left),
+            ),
+          ),
+          Positioned(
+            top: 0,
+            left: 0,
+            height: height,
+            width: width,
+            child: GestureDetector(
+              onTapDown: (_) => WindowUtils.startResize(DragPosition.topLeft),
+            ),
+          ),
+          Positioned(
+            top: 0,
+            right: 0,
+            height: height,
+            width: width,
+            child: GestureDetector(
+              onTapDown: (_) => WindowUtils.startResize(DragPosition.topRight),
+            ),
+          ),
+          Positioned(
+            bottom: 0,
+            left: 0,
+            height: height,
+            width: width,
+            child: GestureDetector(
+              onTapDown: (_) =>
+                  WindowUtils.startResize(DragPosition.bottomLeft),
+            ),
+          ),
+          Positioned(
+            bottom: 0,
+            right: 0,
+            height: height,
+            width: width,
+            child: GestureDetector(
+              onTapDown: (_) =>
+                  WindowUtils.startResize(DragPosition.bottomRight),
+            ),
+          ),
+        ],
       ),
     );
   }